requirements.txt
numpy>=1.26.0
chromadb>=0.5.3
pandas>=2.0.0
pypdf>=5.1.0
openai>=1.60.0
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
umap-learn>=0.5.0
matplotlib>=3.9.0
transformers>=4.0.0
torch>=2.0.0
langchain>=0.3.25
langchain-google-genai>=2.1.5
langgraph>=0.4.8
langsmith>=0.3.45
notebook>=7.4.3
python-dotenv>=1.1.0
scikit-learn>=1.4.0
scipy>=1.11.0
langchain-openai>=0.1.8
langchain-community>=0.2.1
langchain-anthropic>=0.1.15
langchain-google-firestore>=0.3.0
firestore>=0.0.8
tiktoken>=0.7.0
sentence-transformers>=3.0.0
bs4>=0.0.2
firecrawl-py>=0.0.13
langchainhub>=0.1.18
wikipedia>=1.4.0
tavily-python>=0.3.3
streamlit==1.47.1
google_search_results==2.4.2
